+++ client/src/App.tsx
@@ -1,5 +1,5 @@
 import MainHeader from "./components/header/MainHeader";
-import AcheivementsContainer from "./components/content/AcheivementsContainer";
+import AcheivementsContainer from "./components/AcheivementsContainer";
 // import ThemeToggleButton from './components/ThemeToggleButton';
 import { ColorModeButton } from "@/components/ui/color-mode";
 import { HStack, Flex } from "@chakra-ui/react";
